BuildDetails moved to its own class
